Move ChatList helpers into own file Make ChatList a functional component
